# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Platform Enterprise API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/payments/project/{projectId}:
    get:
      operationId: listPaymentScheduleForProject
      tags: [payments]
      summary: List a project's payment schedule (admin)
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment schedule rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiPaymentSchedule"
  /ai/payments/{scheduleId}/verify:
    post:
      operationId: verifyPaymentSchedule
      tags: [payments]
      summary: Verify a submitted payment installment (admin)
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResult"
        "404":
          description: Not found or already verified
...
  /ai/payments/{scheduleId}/reject:
    post:
      operationId: rejectPaymentSchedule
      tags: [payments]
      summary: Reject a submitted payment proof (admin) — P0-5
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRejectInput"
      responses:
        "200":
          description: Payment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPaymentSchedule"
        "404":
          description: Not found or already in terminal state
  /ai/payments/project/{projectId}/unlock:
    post:
      operationId: manualUnlockProjectFiles
      tags: [payments]
      summary: Manually unlock a project's final files (admin override) — P0-5
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualUnlockInput"
      responses:
        "200":
          description: Files unlocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  projectId:
                    type: integer
                  filesUnlocked:
                    type: boolean
                  unlockedBy:
                    type: string
...
  /ai/payments/{scheduleId}/invoice:
    post:
      operationId: generateInvoiceForPaymentSchedule
      tags: [payments]
      summary: Generate an invoice for a payment installment (admin)
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiInvoice"
        "404":
          description: Not found
  /ai/payments/invoices/project/{projectId}:
    get:
      operationId: listInvoicesForProject
      tags: [payments]
      summary: List a project's invoices (admin)
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiInvoice"
  /public/payments/{scheduleId}/submit-proof:
    post:
      operationId: submitPaymentProof
      tags: [payments]
      summary: Customer submits a payment reference for an installment
      parameters:
        - in: path
          name: scheduleId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProofInput"
      responses:
        "200":
          description: Proof recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  schedule:
                    $ref: "#/components/schemas/AiPaymentSchedule"
        "404":
          description: Not found or already paid
...
components:
  schemas:
...
    CreativeProject:
      type: object
      required:
        - id
        - projectId
        - brandName
        - businessType
        - targetMarket
        - productOrService
        - goal
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectId:
          type: string
          description: UUID string — use this as the client-facing ID
        brandName:
          type: string
        businessType:
          type: string
        targetMarket:
          type: string
        productOrService:
          type: string
        stylePreference:
          type:
            - string
            - "null"
        goal:
          type: string
        notes:
          type:
            - string
            - "null"
        status:
          type: string
          description: >
            Legacy: pending | running | completed | failed. Dual Commercial Flow adds:
            waiting_payment | deposit_paid | waiting_payment_verification | payment_verified |
            waiting_remaining_payment | remaining_paid | ready_to_build | building |
            internal_review | waiting_client_review | revision | approved | completed
        paymentPolicy:
          type: string
        depositPercentage:
          type: integer
        paymentStatus:
          type: string
        filesUnlocked:
          type: boolean
        result:
          type:
            - object
            - "null"
          description: Aggregated final output from all agents
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AiPaymentSchedule:
      type: object
      required:
        - id
        - projectId
        - paymentType
        - amount
        - currency
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectId:
          type: integer
        paymentType:
          type: string
          description: deposit | remaining_balance | full_payment | custom_installment | subscription_charge
        percentage:
          type: [integer, "null"]
        amount:
          type: string
        currency:
          type: string
        dueDate:
          type: [string, "null"]
          format: date-time
        status:
          type: string
          description: pending | paid | partially_paid | failed | refunded | cancelled
        reference:
          type: [string, "null"]
        verifiedBy:
          type: [string, "null"]
        paidAt:
          type: [string, "null"]
          format: date-time
        notes:
          type: [string, "null"]
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AiInvoice:
      type: object
      required:
        - id
        - invoiceNumber
        - projectId
        - invoiceType
        - amount
        - currency
        - status
        - issuedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        projectId:
          type: integer
        paymentScheduleId:
          type: [integer, "null"]
        invoiceType:
          type: string
          description: deposit | remaining | final | credit_note | receipt
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
          description: draft | issued | paid | void
        lineItemsJson:
          type: [array, "null"]
          items:
            type: object
            properties:
              label:
                type: string
              amount:
                type: number
        issuedAt:
          type: string
          format: date-time
        paidAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaymentRejectInput:
      type: object
      required: [rejectedBy, reason]
      properties:
        rejectedBy:
          type: string
        reason:
          type: string
    ManualUnlockInput:
      type: object
      required: [unlockedBy]
      properties:
        unlockedBy:
          type: string
        reason:
          type: string
...
    PaymentVerifyInput:
      type: object
      required:
        - verifiedBy
      properties:
        verifiedBy:
          type: string
        reference:
          type: string
    PaymentVerifyResult:
      type: object
      required:
        - schedule
        - project
        - productionStarted
      properties:
        schedule:
          $ref: "#/components/schemas/AiPaymentSchedule"
        project:
          $ref: "#/components/schemas/CreativeProject"
        productionStarted:
          type: boolean
    PaymentProofInput:
      type: object
      required:
        - reference
      properties:
        reference:
          type: string
...